requirements.txt
streamlit
boto3==1.38.30
requests
awscli
streamlit-webrtc
pydub
fastapi
uvicorn
aws-requests-auth
python-multipart
websockets==10.4
amazon-transcribe
python-dotenv
invoke